chore: expose SortableHeader
- SortableHeader was within InvoicesTable and could be used generically, so pull it out
- add stories to demo
